+++ src/Graphics/UI/Threepenny/Internal/Driver.hs
@@ -40,12 +40,12 @@ module Graphics.UI.Threepenny.Internal.Driver
   
   -- * Utilities
   ,debug
-  ,callDeferredFunction
   ,atomic
 
   -- * JavaScript FFI
   ,ToJS, FFI, ffi, JSFunction
   ,runFunction, callFunction
+  ,newHsFunction
   
   -- * Types
   ,Window
@@ -347,17 +347,6 @@ run :: Session -> Instruction -> IO ()
 run (Session{..}) instruction =
     writeChan sInstructions $!! instruction  -- see note [Instruction strictness]
 
--- | Call the given function with the given continuation. Doesn't block.
-callDeferredFunction
-  :: Window                    -- ^ Browser window
-  -> String                    -- ^ The function name.
-  -> [String]                  -- ^ Parameters.
-  -> ([Maybe String] -> IO ()) -- ^ The continuation to call if/when the function completes.
-  -> IO ()
-callDeferredFunction window fun params thunk = do
-    closure <- newClosure window fun $ \(EventData xs) -> thunk xs
-    run window $ CallDeferredFunction (closure,fun,params)
-
 -- | Run the given JavaScript function and carry on. Doesn't block.
 --
 -- The client window uses JavaScript's @eval()@ function to run the code.
@@ -376,21 +365,20 @@ callFunction window jsfunction =
                 Error   _ -> return Nothing
             _ -> return Nothing
 
-
 -- | Package a Haskell function such that it can be called from JavaScript.
--- 
+--
 -- At the moment, we implement this as an event handler that is
--- attached to the @head@ element.
-newClosure
-    :: Window               -- ^ Browser window context
-    -> String               -- ^ Function name (for debugging).
-    -> (EventData -> IO ()) -- ^ Function to call
-    -> IO Closure
-newClosure window@(Session{..}) fun thunk = do
+-- attached to the @head@ element. In particular, it is not garbage
+-- collected as long as the head element is alive.
+newHsFunction
+    :: Window       -- ^ Browser window context
+    -> IO ()        -- ^ Haskell function
+    -> IO (HsFunction (IO ()))
+newHsFunction window@(Session{..}) fun = do
     cid <- modifyMVar sClosures $ \(x:xs) -> return (xs,x)
-    let eventId = fun ++ "-" ++ show cid
-    attachClosure sHeadElement eventId thunk
-    return $ Closure (unprotectedGetElementId sHeadElement, eventId)
+    let eventId = "callback-" ++ show cid
+    attachClosure sHeadElement eventId (const $ fun)
+    return $ HsFunction (unprotectedGetElementId sHeadElement) eventId
 
 
 {-----------------------------------------------------------------------------