+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog for the `threepenny-gui` package
 
+**0.5.0.0** -- Snapshot release.
+
+* FFI now supports callbacks into Haskell. Remove `callDeferredFunction` function.
+* Remove dependency on `json` library.
+
+**0.4.0.1** -- Maintenance release. Adjust package dependencies.
+
 **0.4.0.0** -- Snapshot release.
 
 * New `UI` monad for easier JavaScript FFI and recursion in FRP.